remove service arguments
reason: could not determine why the argument was necessary.

The original reason for adding -u is given in
xivo-websocketd:5ca679893e879c7605004d5f47c0f4a974497be4, to allow
stdout to be logged in journalctl.  I did not see any difference
between -u and not -u, so I chose the simplest of the two