participants: use subqueryload
reason: joinedload in this context generates a triple OUTER JOIN on the
same table in a single SQL query. Given 1000 rows in
call_log_participant, the triple OUTER JOIN creates a 1000^3 rows
result, eating up all available RAM on the PG client side (psql or
sqlalchemy). With subqueryload, the query is split in 3 and
source/destination participants are queried separately.