push mobile: avoid generating extra call_logs
this solution is a quick fix to avoid generating multiple call_logs for a single
call. The cels matching the skipped call_log will remain with no call_log_id and
will be analyzed again and again and again.

again and again, again and again.

which is hurtful in the long run