Revert "Merge branch 'push_mobile'"
This reverts commit 6f8995e8cb70afad88c82b89eaca9c525593eeac, reversing
changes made to 57d2f238c2aad6833f04d15c7acc1824105e92cb.

wazo-calld now assign the originator when calling the webrtc endpoints, this
result in a single linkedid being used for all the legs of the call.

the code that ignored the CELS of a dial_mobile is now useless since it was
based on the assumption that many linkedids would exist for a single call.