allow to not specify key_file
reason: when installing call-logd, postinst run alembic and the config is
loaded. But keys are generated only in wazo-upgrade pre-start. So we
need to authorize loading config without key_file. Moreover, if the
user set a custom config without key_file but with username/password
directly in the configuration, we don't want to raise an error.