cdr: add the recording filename
when listing cdr add the recording filename to the result.

this commit also changes the filename the first part of the name is now the
start time of the recording instead of the start time of the call.

this change makes the file sortable in order when listening to all files of a
call.