cdr: allow search CDR by recording filename
If the searched term matches the filename format for a recording the search will
be filtered based on the call log ID and recording UUID in the searched
filename. The start date and extension are ignored.

This will always work with real filenames and may return weird results for users
forging fake filenames with a valid cdr id and recording uuid.