debian: pin sqlalchemy-utils to fix upgrade traceback
why: When upgrading from Buster to Bullseye, sqlalchemy-utils is
upgraded after wazo services. So if the service use new method, it will
raise an error and fail the installation of this package.

Fortunally, on most case, the command after is dist-upgrade and
will fix thing. But it's still worrying for a sysadmin